+++ reserva-app/database/migrations/2025_11_24_194801_add_uuid_to_users_table.php
@@ -11,9 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->uuid('uuid')->unique()->after('id');
-        });
+        if (env('DB_CONNECTION') == 'testing_memory') {
+            // recriado em outras migrations
+            return;
+        }
+
+        if (env('DB_CONNECTION') != 'testing_memory') {
+            Schema::table('users', function (Blueprint $table) {
+                $table->uuid('uuid')->unique()->after('id');
+            });
+        }
     }
 
     /**
@@ -21,8 +28,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('uuid');
-        });
+        if (env('DB_CONNECTION') != 'testing_memory') {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('uuid');
+            });
+        }
     }
 };
