+++ reserva-app/database/migrations/2025_11_24_205821_add_status_to_users_table.php
@@ -11,9 +11,24 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('status')->default('active')->after('role');
-        });
+        if (env('DB_CONNECTION') == 'testing_memory') {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('name')->nullable(false);
+                $table->string('email')->unique();
+                $table->string('password');
+                $table->enum('role', ['recepcionista', 'admin', 'gerente', 'garcom'])->default('recepcionista');
+                $table->enum('status', ['active', 'inactive'])->default('active');
+                $table->timestamps();
+            });
+            return;
+        }
+
+        if (env('DB_CONNECTION') != 'testing_memory') {
+            Schema::table('users', function (Blueprint $table) {
+                $table->enum('status', ['active', 'inactive'])->default('active')->after('role');
+            });
+        }
     }
 
     /**
@@ -21,8 +36,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('status');
-        });
+        if (env('DB_CONNECTION') != 'testing_memory') {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('status');
+            });
+        }
     }
 };
